Remove "containers" directory, move files in context w/in views
