+++ airflow/ui/src/views/Config/Pools.tsx
@@ -20,7 +20,7 @@
 import React from 'react';
 import { Heading } from '@chakra-ui/react';
 
-import ConfigContainer from 'containers/ConfigContainer';
+import ConfigContainer from './ConfigContainer';
 
 const Pools: React.FC = () => (
   <ConfigContainer current="Pools">
